Safari-609.1.11.1/LayoutTests/http/tests/security/contentSecurityPolicy/connect-src-websocket-blocked-expected.txt

CONSOLE MESSAGE: Refused to connect to ws://localhost:8880/websocket/tests/hybi/echo because it does not appear in the connect-src directive of the Content Security Policy.
Pass
